Added reset button functionality;

+++ index.php
@@ -6,18 +6,27 @@
         $_SESSION['activeList'] = array();
         $_SESSION['completedList'] = array();
     }
-    var_dump($_SESSION['activeList']);
-    $newToDo = FALSE;
+    $newToDo = '';
 
     include './templates/header.php';
 
-
-    if ( ( !empty( $_POST)) && ($_POST['newTask'] !== '' ))
+    if(isset($_POST['add']))
     {
-        $newToDo = $_POST['newTask'];
-        array_push( $_SESSION[ 'activeList' ], "{$newToDo}" );
-        $newToDo = '';
+        if ( ( !empty( $_POST)) && ($_POST['newTask'] !== '' ) && ( !in_array( $_POST['newTask'], $_SESSION['activeList']) ) )
+        {
+            $newToDo = $_POST['newTask'];
+            array_push( $_SESSION[ 'activeList' ], "{$newToDo}" );
+            $newToDo = '';
+        }
+        var_dump($_SESSION['activeList']);
     }
+    if(isset($_POST['reset']))
+    {
+        session_unset();
+        session_destroy();
+    }
+
+    
 ?>
 
 <h1>Welcome to Krisztina's PHP To-Do List</h1>
@@ -33,17 +42,22 @@
             type="string"
             value="">
     </label>
-    <input type="submit" value="Add To List">
+    <input type="submit" name="add" value="Add To List">
+    <input type="submit" name="reset" value="Reset List">
 </form>
 
 <h2>Active To-Dos</h2>
 <ul>
-<?php foreach ( $_SESSION[ 'activeList' ] as $task ) : ?>
-    <li>
-        <?php echo $task ?>
-        <button type="delete" value="delete">Delete</button>
-    </li>
-<?php endforeach; ?>
+<?php if ( !empty($_SESSION['activeList']) ) : ?> 
+    <?php foreach ( $_SESSION[ 'activeList' ] as $task ) : ?>
+        <li>
+            <?php echo $task ?>
+            <button type="delete" value="delete">Delete</button>
+        </li>
+    <?php endforeach; ?>
+<?php endif ?>
+
+
 </ul>
 
 <h2>Completed To-Dos</h2>
